Don't underline dropdown links when hovering
Bootstrap 5 underlines all links by default. We don't want that
feature, so we've set text-decoration: none to links by default
and text-decoration: underline when hovering (as in Bootstrap 3).
However, globally enabling underline when hovering for every link
overrides certain occasions which Bootstrap knows should not have
underlining, such as items in dropdown. Therefore, we must disable
it manually for cases such as those.